chore: split two long functions to clear noExcessiveLinesPerFunction warnings
Two Biome warnings I had been deferring with "pre-existing" framing.
They are not pre-existing in any meaningful sense — they are real
warnings, the user sees them every time CI runs, and we've decided
warnings are latent failures that need fixing (see PR #136).

Splits:

- frontend/features/knowledge/KnowledgeContainer.tsx (was 178 lines):
  extracted URL+data state into useKnowledgeUrlState() and the
  navigation handler suite into useKnowledgeNavigation().  The
  container now reads as render orchestration only — if/else for
  loading/error states, then renders KnowledgeView.  Each new hook
  is independently testable.

- frontend/features/whimsy/index.tsx::WhimsySettingsCard (was 182
  lines): extracted three sub-components inside the same module —
  WhimsyPresetModeRows, WhimsyThemeRow, WhimsyGeneratedModeRows —
  each rendering one branch of the mode switch.  No new files (the
  rows are tightly coupled to whimsy-local types and bounds; pulling
  them out further would just spread the import surface).

Verification (local):

  bunx --bun @biomejs/biome check    -> 0 errors, 0 warnings
  bun x tsc --noEmit                  -> green
  node ../scripts/check-file-lines.mjs -> OK
  node ../scripts/check-nesting.mjs   -> OK

No behaviour changes — the extracted hooks return the same shapes the
inline code did, and the row components consume the same config +
setConfig pair via a single WhimsyRowProps type.
